backend/requirements.txt
aiofiles
aiohttp
alembic
arxiv
asyncio
asyncpg
beautifulsoup4
bs4
colorama
duckduckgo_search
fastapi
google-ads
gunicorn
htmldocx
httptools
jinja2
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
openai
passlib[bcrypt]
playwright
pydantic
pydantic-settings
PyMuPDF
python-dotenv
python-jose[cryptography]
python-multipart
pyjwt
pyyaml
redis
requests
rich
selenium
sqlalchemy
tavily-python
toml
uvicorn
uvloop
websockets
cloud-sql-python-connector
pg8000
google-search-results
langchainhub